Final steps of the MVC Movie Store tutorial

+++ MvcMovie2/Controllers/MovieController.cs
@@ -103,9 +103,13 @@ namespace MvcMovie2.Controllers
         // POST: /Movie/Delete/5
 
         [HttpPost, ActionName("Delete")]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int id = 0)
         {            
             Movie movie = db.Movies.Find(id);
+
+            if (movie == null)
+                return HttpNotFound();
+
             db.Movies.Remove(movie);
             db.SaveChanges();
             return RedirectToAction("Index");
